requirements.txt
# transformers==4.26.1
# torch==1.13.1
transformers==4.44.1
torch==2.4.1
Pillow==10.4.0
accelerate==0.34.2
scipy==1.14.1
numpy<2